+++ DataObject/StoreInventory.cs
@@ -1,14 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Text;
 
 namespace DataObject
 {
     public class StoreInventory
     {
-        public long StoreID { get; set; }
-        public long ProductID { get; set; }
-        public long StockLevel { get; set; }
+        //[Key]
+        public int StoreID { get; set; }
+        //[Key]
+        public int ProductID { get; set; }
+        public int StockLevel { get; set; }
 
         //expanded from product
         public string ProductName { get; set; }
